Gestures Part 1

+++ components/FavoriteComponent.js
@@ -1,11 +1,11 @@
 import React, { Component } from 'react';
-import { FlatList, View, Text } from 'react-native';
+import { FlatList, View, Text, Alert } from 'react-native';
 import { ListItem } from 'react-native-elements';
 import { connect } from 'react-redux';
 import { Loading } from './LoadingComponent';
 import { baseUrl } from '../shared/baseUrl';
-import Swipeout from 'react-native-swipeout';
 import { deleteFavorite } from '../redux/ActionCreators';
+import * as Animatable from 'react-native-animatable';
 
 const mapDispatchToProps = dispatch => ({
     deleteFavorite: (dishId) => dispatch(deleteFavorite(dishId))
@@ -36,12 +36,30 @@ class Favorites extends Component {
                 {
                     text: 'Delete', 
                     type: 'delete',
-                    onPress: () => this.props.deleteFavorite(item.id)
+                    onPress: () => {
+                        Alert.alert(
+                            'Delete Favorite?',
+                            'Are you sure you wish to delete the favorite dish ' + item.name + '?',
+                            [
+                                { 
+                                    text: 'Cancel', 
+                                    onPress: () => console.log(item.name + 'Not Deleted'),
+                                    style: ' cancel'
+                                },
+                                {
+                                    text: 'OK',
+                                    onPress: () => this.props.deleteFavorite(item.id)
+                                }
+                            ],
+                            { cancelable: false }
+                        );
+                        
+                    }
                 }
             ];
     
             return (
-                <Swipeout right={rightButton} autoClose={true}>
+                <Animatable.View animation="fadeInRightBig" duration={2000}>                
                     <ListItem
                         key={index}
                         title={item.name}
@@ -50,7 +68,7 @@ class Favorites extends Component {
                         onPress={() => navigate('Dishdetail', { dishId: item.id })}
                         leftAvatar={{ source: {uri: baseUrl + item.image}}}
                         />
-                </Swipeout>
+                </Animatable.View>
             );
         };
